% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{burp}
\alias{burp}
\title{Select a color palette}
\usage{
burp(
  name,
  pal_class = c("Hero", "Villain", "Place"),
  n,
  type = c("discrete", "continuous")
)
}
\arguments{
\item{name}{A string corresponding to one of the available palette.}

\item{pal_class}{A string between Hero, Villain or Place}

\item{n}{A number of color}

\item{type}{A string}
}
\value{
A color palette in a form of a vector
}
\description{
Select a color palette
}
\examples{
burp("Shrek", pal_class="Hero", n=5)
burp("Lord_Farquaad", pal_class="Villain", n=6)
}
