% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{hero_palettes}
\alias{hero_palettes}
\title{Heroes' palettes}
\format{
An object of class \code{list} of length 6.
}
\usage{
hero_palettes
}
\description{
List of palettes inspired from Heroes
}
\keyword{datasets}
