% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{place_palettes}
\alias{place_palettes}
\title{Places' palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
place_palettes
}
\description{
list of palettes inspired from Places
}
\keyword{datasets}
