% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{show_my_swamp}
\alias{show_my_swamp}
\title{Plot a color palette}
\usage{
show_my_swamp(pal)
}
\arguments{
\item{pal}{A color palette}
}
\value{
A ggplot objects
}
\description{
Plot a color palette
}
\examples{
shrek_pal = burp("Shrek", pal_class="Hero", n=6)
show_my_swamp(shrek_pal)
}
