% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{villain_palettes}
\alias{villain_palettes}
\title{Villains' palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
villain_palettes
}
\description{
list of palettes inspired from Villain
}
\keyword{datasets}
